Performance improvements
These changes reduce the average run time against a sample log from about
100 seconds to about 60 seconds on my laptop, with only two small changes.

First, I removed the Add-ServerPerformanceObject_ValueTemp function. This
function took an Array and returned a List<Object>, which has no value
here. The only reason we would want to convert to a List is if we were
going to manipulate this data - add items, remove items, etc. But this is
a static list of performance counter samples read from a file. We won't be
adding and removing anything, so this conversion has no value. This change
shaved 90% off the time off the Convert- function.

Second, I removed the Select-Object -Skip 1 projection in
Add-ServerPerformanceObject_Value. This was projecting a whole new
sequence of a large list just so it didn't contain the first item in the
list. This change cuts about 30% of the time off the Analyze- function.

Note this second change means we are now including the first sample in the
analysis. This is a change in behavior from the original code. If we
really do need to skip that first sample, we should find a more efficient
way to do it.