Performance improvements part 2
This change removes the FileName attribute, providing a significant
performance boost. This cuts average run time in half again on top of the
first set of performance improvements.

The code allows you to pass many perf files at once. To keep track of which
sample came from which file, we were performing an expensive Add-Member
operation on every individual sample. This didn't do much for us, though,
because we didn't use these anywhere except to set a value on the server
object. And that value was meaningless, since we could have multiple
files, but we only stored one file name on the server.

Removing that expensive Add-Member operation cuts the
Get-PerformanceDataFromFileLocal time from about 20 seconds to about 13
seconds on my laptop, shaving 35% off that time.

More surprisingly, removing that FileName attribute also makes
Analyze-DataOfObject dramatically faster, and I'm not sure why. On my
laptop, running the code from the first round of perf improvements, it
takes Analyze-DataOfObject about 25 seconds on the sample perflog. If we
remove FileName, Analyze-DataOfObject takes only 7 seconds.

This means that a run time of 60 seconds against the sample data is cut in
half to only 30 seconds with this small change.