Removed remaining Add-Member calls
This change removes the last calls to Add-Member for consistency. It
doesn't make sense to define a model in C# and then call Add-Member to add
a dynamic property to a model that we control. Just define the member on
the model.

These last few Add-Member calls weren't called often enough to noticably
affect performance.