Skip the first sample
Per David, we do need to skip the first sample because it is sometimes
inaccurate. In the first round of perf changes, I took out the expensive
Select-Object projection that was skipping the first sample.

This change reintroduces the skip first behavior by using Linq. We now
assign the object[] to AllRawData, but when we iterate over the data use
use RawData, which is now an IEnumerable that performs a Skip(1).

I also broke the averages code in a previous change, which I fixed here.